requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==20.3.0
cachelib==0.10.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.2
cs50==9.2.5
distlib==0.3.6
filelock==3.9.0
Flask==2.2.3
Flask-Session==0.4.0
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
jellyfish==0.9.0
Jinja2==3.1.2
lib50==3.0.5
MarkupSafe==2.1.2
multidict==6.0.4
openai==0.27.0
pbr==5.11.1
pexpect==4.8.0
pip==23.0.1
platformdirs==3.1.0
ptyprocess==0.7.0
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.17
python-dotenv==1.0.0
PyYAML==5.4.1
requests==2.28.2
setuptools==65.5.0
SQLAlchemy==1.4.46
sqlparse==0.4.3
stevedore==5.0.0
termcolor==1.1.0
tqdm==4.65.0
urllib3==1.26.14
virtualenv==20.20.0
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
Werkzeug==2.2.3
wheel==0.38.4
yarl==1.8.2
